requirements.txt
alembic==1.7.7
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
